Add mouse gesture manager
I do this to make it easier to support different input modes.

The idea is that we can listen to different kind of gestures and have
a callback called at the start, middle and end of a gesture.  The
gesture manager (gesture.c) takes care of deciding when a gesture starts
and end.