Some minor cleanup for new gesture code
No need to pass the 'inside' argument anymore.  If we want to check that
the mouse if really inside the view we can always compute it.

Pre-computing this doesn't make sense on a touch screen interface where
we could have some of the fingers inside and other outside the view.