Put block voxels origin at the corner
Instead of at the middle, that is, in a block at position (x, y, z),
the pixel at index (0, 0, 0) is at the position (x, y, z) (before this
commit it was at (x - N/2, y - N/2, x - N/2)).

I should have done like this from the start!  There is a lot of tricky
parts in the code because of this choice.  For the moment I just change
the code so that it compiles and run, but many things are probably
broken, like the file format.