Disable travis gcc build for the moment
gcc 4 annoyingly follows the C standard and refuses to call a function
with a const multidimensional array argument with a non const array.

Fortunately gcc 5 only issue this error in pedantic mode (and even then
we can disable it with a flag).

Until travis support gcc 5, I am just going to disable the gcc build.  I
guess clang is enough.