Remove goxel->pick_mesh
It was used so that we can edit the current layer mesh without affecting
the snapping.  Instead we now have to make sure no tool edit the actual
layer mesh until the drawing operation is over.  In order to preview a
drawing operation we can use goxel->tool_mesh (the name might need to be
changed to reflect the more general usage).