Optimize mesh generate vertice function
Unfortunately mesh_get_at is really too slow as it is implemented now!

I tried to optimize it but without success.  For the moment we copy the
block data out of the mesh so that we can access it directly.  Much
faster!

Some other part of the program are still too slow though.