mesh: keep key in add_neighboor_blocks
This function is already a hack, at some point I should try to delete
it.  The code rely on the key not changing when no operation is done on
a mesh, so we have to make sure that we only change the key when
necessary.

Eventually I would like to compute the key using an actual md5 or crc64
algo.