Big optimization of obj/ply export
We don't search the whole list for each entry.  I think we could maybe
use a hash table here for better optimization, but I don't want to work
too much on this since eventually we also need to support simplified
mesh export with merged faces, and the best would be to use an external
obj generation library if it exists.

At least now the export of big models takes seconds instead of hours!