Image use key to know if it changed
I do that so that I can disable the save menu if the image hasn't been
changed.

This is cleaner than having a 'dirty' attribute I think, but I am a bit
worried this might be a bit slow since it forces us to recompute the key
at each iteration.

If this pose problem in the future I'll remove it.