Make the global goxel instance not a pointer
I do that because I cannot find a clean way to let Swift access
attributes of pointer to C structure!  There are several ways to expose
the goxel attributes to Swift, but the simplest seems to make the global
instance a struct instead of a pointer to it.

Even though it's a big commit, It shouldn't change anything, I just
replace all the 'goxel->' to 'goxel.' The compiled code should stay the
same I think.  Probably breaks the osx build though.